move multi-file-list/reader to a smaller scope
